+++ src/Api/Cups/Attributes/UnsupportedGroup.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Rawilk\Printing\Api\Cups\Attributes;
 
 use Rawilk\Printing\Api\Cups\AttributeGroup;
-use Rawilk\Printing\Api\Cups\AttributeGroupTag;
+use Rawilk\Printing\Api\Cups\Enums\AttributeGroupTag;
 
 class UnsupportedGroup extends AttributeGroup
 {
